ensure scale transform return the double values.
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/flowWorkspace@116306 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/transformation.cpp
@@ -447,7 +447,7 @@ void linTrans::transforming(valarray<double> & input){
 
 void scaleTrans::transforming(valarray<double> & input){
 
-		input*=(t_scale/r_scale);
+		input*=(t_scale/(double)r_scale);
 }
 
 void flinTrans::transforming(valarray<double> & input){
